+++ packages/webcodecs/src/on-video-track.ts
@@ -2,9 +2,7 @@ import type {
 	MediaParserLogLevel,
 	MediaParserOnVideoTrack,
 } from '@remotion/media-parser';
-import {arrayBufferToUint8Array} from './arraybuffer-to-uint8-array';
 import {canCopyVideoTrack} from './can-copy-video-track';
-import {convertEncodedChunk} from './convert-encoded-chunk';
 import type {ConvertMediaOnVideoFrame} from './convert-media';
 import type {MediaFn} from './create/media-fn';
 import type {ProgressTracker} from './create/progress-tracker';
@@ -13,16 +11,10 @@ import type {ConvertMediaContainer} from './get-available-containers';
 import type {ConvertMediaVideoCodec} from './get-available-video-codecs';
 import {getDefaultVideoCodec} from './get-default-video-codec';
 import {Log} from './log';
-import {onFrame} from './on-frame';
 import type {ConvertMediaOnVideoTrackHandler} from './on-video-track-handler';
-import {processingQueue} from './processing-queue';
+import {reencodeVideoTrack} from './reencode-video-track';
 import type {ResizeOperation} from './resizing/mode';
-import {calculateNewDimensionsFromRotateAndScale} from './rotation';
 import type {ConvertMediaProgressFn} from './throttled-state-update';
-import {createVideoDecoder} from './video-decoder';
-import {getVideoDecoderConfigWithHardwareAcceleration} from './video-decoder-config';
-import {createVideoEncoder} from './video-encoder';
-import {getVideoEncoderConfig} from './video-encoder-config';
 import type {WebCodecsController} from './webcodecs-controller';
 
 export const makeVideoTrackHandler =
@@ -121,183 +113,16 @@ export const makeVideoTrackHandler =
 			};
 		}
 
-		if (videoOperation.type !== 'reencode') {
-			throw new Error(
-				`Video track with ID ${track.trackId} could not be resolved with a valid operation. Received ${JSON.stringify(
-					videoOperation,
-				)}, but must be either "copy", "reencode", "drop" or "fail"`,
-			);
-		}
-
-		const rotation = (videoOperation.rotate ?? rotate) - track.rotation;
-
-		const {height: newHeight, width: newWidth} =
-			calculateNewDimensionsFromRotateAndScale({
-				width: track.codedWidth,
-				height: track.codedHeight,
-				rotation,
-				videoCodec: videoOperation.videoCodec,
-				resizeOperation: videoOperation.resize ?? null,
-			});
-
-		const videoEncoderConfig = await getVideoEncoderConfig({
-			codec: videoOperation.videoCodec,
-			height: newHeight,
-			width: newWidth,
-			fps: track.fps,
-		});
-		const videoDecoderConfig =
-			await getVideoDecoderConfigWithHardwareAcceleration(track);
-
-		Log.verbose(logLevel, 'Video encoder config', videoEncoderConfig);
-		Log.verbose(logLevel, 'Video decoder config', videoDecoderConfig ?? track);
-
-		if (videoEncoderConfig === null) {
-			abortConversion(
-				new Error(
-					`Could not configure video encoder of track ${track.trackId}`,
-				),
-			);
-			return null;
-		}
-
-		if (videoDecoderConfig === null) {
-			abortConversion(
-				new Error(
-					`Could not configure video decoder of track ${track.trackId}`,
-				),
-			);
-			return null;
-		}
-
-		const {trackNumber} = await state.addTrack({
-			type: 'video',
-			color: track.advancedColor,
-			width: newWidth,
-			height: newHeight,
-			codec: videoOperation.videoCodec,
-			codecPrivate: null,
-			timescale: track.originalTimescale,
-		});
-		Log.verbose(
-			logLevel,
-			`Created new video track with ID ${trackNumber}, codec ${videoOperation.videoCodec} and timescale ${track.originalTimescale}`,
-		);
-
-		const videoEncoder = createVideoEncoder({
-			onChunk: async (chunk, metadata) => {
-				await state.addSample({
-					chunk: convertEncodedChunk(chunk),
-					trackNumber,
-					isVideo: true,
-					codecPrivate: arrayBufferToUint8Array(
-						(metadata?.decoderConfig?.description ??
-							null) as ArrayBuffer | null,
-					),
-				});
-				onMediaStateUpdate?.((prevState) => {
-					return {
-						...prevState,
-						encodedVideoFrames: prevState.encodedVideoFrames + 1,
-					};
-				});
-			},
-			onError: (err) => {
-				abortConversion(
-					new Error(
-						`Video encoder of track ${track.trackId} failed (see .cause of this error)`,
-						{
-							cause: err,
-						},
-					),
-				);
-			},
+		return reencodeVideoTrack({
+			videoOperation,
+			abortConversion,
 			controller,
-			config: videoEncoderConfig,
 			logLevel,
-			outputCodec: videoOperation.videoCodec,
+			rotate,
+			track,
+			onVideoFrame,
 			progress,
+			state,
+			onMediaStateUpdate,
 		});
-
-		const videoProcessingQueue = processingQueue<VideoFrame>({
-			controller,
-			label: 'VideoFrame processing queue',
-			logLevel,
-			onError: (err) => {
-				abortConversion(
-					new Error(
-						`VideoFrame processing queue of track ${track.trackId} failed (see .cause of this error)`,
-						{
-							cause: err,
-						},
-					),
-				);
-			},
-			onOutput: async (frame) => {
-				await onFrame({
-					frame,
-					track,
-					videoEncoder,
-					onVideoFrame,
-					outputCodec: videoOperation.videoCodec,
-					rotation,
-					resizeOperation: videoOperation.resize ?? null,
-				});
-			},
-		});
-
-		const videoDecoder = createVideoDecoder({
-			config: videoDecoderConfig,
-			onFrame: async (frame) => {
-				await videoProcessingQueue.ioSynchronizer.waitForQueueSize({
-					queueSize: 10,
-					controller,
-				});
-				videoProcessingQueue.input(frame);
-			},
-			onError: (err) => {
-				abortConversion(
-					new Error(
-						`Video decoder of track ${track.trackId} failed (see .cause of this error)`,
-						{
-							cause: err,
-						},
-					),
-				);
-			},
-			controller,
-			logLevel,
-		});
-
-		state.addWaitForFinishPromise(async () => {
-			Log.verbose(logLevel, 'Waiting for video decoder to finish');
-			await videoDecoder.waitForFinish();
-			videoDecoder.close();
-			Log.verbose(
-				logLevel,
-				'Video decoder finished. Waiting for encoder to finish',
-			);
-			await videoEncoder.waitForFinish();
-			videoEncoder.close();
-			Log.verbose(logLevel, 'Encoder finished');
-		});
-
-		return async (chunk) => {
-			progress.setPossibleLowestTimestamp(chunk.timestamp);
-			await progress.waitForMinimumProgress({
-				minimumProgress: chunk.timestamp - 10_000_000,
-				controller,
-			});
-
-			await videoDecoder.ioSynchronizer.waitForQueueSize({
-				queueSize: 20,
-				controller,
-			});
-
-			if (chunk.type === 'key') {
-				await videoDecoder.flush();
-			}
-
-			videoDecoder.decode(chunk);
-		};
 	};
